Use double word for access vector
due to alignment, it doesn't actually take any more space (x86-64)

+++ ruby/bop/object_monitor.h
@@ -27,7 +27,7 @@ typedef struct{
     };
     volatile record_id_t record_id;
   };
-  volatile uint64_t vector;
+  volatile uint128_t vector;
 #ifdef HAVE_USE_PROMISE
   bool id_valid;
 #endif
